fix: pq hnsw
